Separate bsp4j from docs codegen
