+++ codegen/src/main/scala/bsp/codegen/bsp4j/Codegen.scala
@@ -1,7 +1,6 @@
-package bsp.codegen
+package bsp.codegen.bsp4j
 
-import bsp.codegen.bsp4j.JavaRenderer
-import bsp.codegen.docs.MarkdownRenderer
+import bsp.codegen.ModelLoader
 import bsp.codegen.ir.SmithyToIR
 
 object Codegen {
@@ -28,15 +27,4 @@ object Codegen {
     }
   }
 
-  def docs(namespace: String): String = {
-    val model = ModelLoader.loadModel()
-    val docTree = new SmithyToIR(model).docTree(namespace)
-    MarkdownRenderer.render(docTree)
-  }
-
-  def printDocs(namespace: String): Unit = println(docs(namespace))
-
-  def main(args: Array[String]): Unit = {
-    printDocs("bsp")
-  }
 }
