+++ docs/extensions/java.md
@@ -8,5 +8,5 @@ The following section contains Java-specific extensions to the build server
 protocol.
 
 ```scala mdoc:passthrough
-bsp.codegen.Codegen.printDocs("bsp.java")
+bsp.codegen.docs.Codegen.printDocs("bsp.java")
 ```
