+++ docs/extensions/jvm.md
@@ -8,5 +8,5 @@ The following section contains JVM-specific extensions to the build server
 protocol.
 
 ```scala mdoc:passthrough
-bsp.codegen.Codegen.printDocs("bsp.jvm")
+bsp.codegen.docs.Codegen.printDocs("bsp.jvm")
 ```
